src/backend/common/jwt.py
"""JWT Authentication Module"""
from datetime import datetime, timedelta
from asgiref.sync import sync_to_async
from fastapi import Depends, Request
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm, HTTPBearer
from fastapi.security.utils import get_authorization_scheme_param
from jose import jwt
from passlib.context import CryptContext
from sqlmodel.ext.asyncio.session import AsyncSession


from backend.common.exceptions.errors import AuthorizationError, TokenExpiredError